LPS-162880 split PublicationsSearchContainer into its own file
